requirements.txt
asgiref==3.11.0
Django==5.2.8
mysqlclient==2.2.7
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.4
tzdata==2025.2
argon2-cffi==23.1.0
django-redis==5.4.0
redis==5.0.1
gunicorn==21.2.0
celery==5.3.6
djangorestframework==3.14.0
drf-spectacular==0.27.1
